Refactor style system for platform-agnostic rendering
Major refactor to decouple spreadsheet style/drawing types from WPF, introducing a cross-platform style model. Added new core types (`CellColor`, `CellFontFamily`, `CellFontWeight`, etc.) and replaced all style-related APIs to use these. Removed WPF-specific style classes and introduced `WpfResourceCache` for mapping core styles to WPF resources. Updated all rendering, editing, and sample code to use the new model. Cleaned up obsolete files and moved sorting/data collection to core. This enables clear separation of core logic and UI, improving portability and maintainability. Added `.editorconfig` for code style.

+++ src/DevBrewLabs.Spreadsheet/Core/Extensions.cs
@@ -22,7 +22,7 @@ namespace DevBrewLabs.Spreadsheet
             while (true)
             {
                 mod = (index % 26) + 65;
-                index = (int)(index / 26);
+                index = index / 26;
                 achar = (char)mod;
                 str = achar + str;
                 if (index > 0) index--;
